requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
crewai==0.28.0
langchain==0.0.312
langchain-openai>=0.0.5
langchain-groq>=0.1.1
openai==1.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
gunicorn==21.2.0
alembic==1.13.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.23.2
black==23.11.0
isort==5.13.0
flake8==6.1.0
tenacity==8.2.3
uvloop==0.19.0
httptools==0.6.1
